Big, Nasty bug
There was a bug in the read_file implementaiton in fsimpl. I assumed
that blocks would be full. While adding a disassembler as a WASM
program, I was having weird problems writing to a file. It turns out
that each write would consume a block (yes, this is bogus). So when
reading bolcks back, my assumption caused the decryption to be
off, and everything got weird.

+++ src/fuse.rs
@@ -641,7 +641,7 @@ impl<B: BlockStorage> Filesystem for UberFSFuse<B> {
         );
 
         let guard = self.file_system.lock().expect("poisoned ufs lock");
-        match &mut guard.read_file(fh, offset as u64, size as usize) {
+        match &mut guard.read_file(fh, offset as u64, size) {
             Ok(buffer) => {
                 debug!("read {} bytes", buffer.len());
                 trace!("{:?}", &buffer);
